Made four lemming fiesta level tiles match
Made the tiles match the easier variant's. Also fiddled around with the
exit distance and pipes a little. This now means that the whole ordeal
is 2 or 3 seconds slower than before (leaving us with under 5 seconds
spare), but as far as i can tell it shouldn't be possible to do the
correct solution and fail on time if you actually saved all the lemmings
in the first place, even if one ends up climbing?