New version of four lemming fiesta
The old level had an unfixable backroute where you could just trap the
lemmings in the top and not use the blocker reversal. Created a new
level which uses the same puzzle element for letting one guy go ahead
and built a new level around it. Still a 1 min time limit though, might
want to increase it if it to 2 mins (or 1:11 in neolemmix) depending on
feedback?